Added "Ping-Pong" game art and assets.
what i simply added was my assets that i used for the game, like a background, a ball, a line, and a paddle, very similarly for my "breakout" game, as please note that these are prototypes. and not final artwork. as i might do something creative im currently having in mind.