Implimented Scene Transitions but still buggy and needs polishing
I implimented scene transitions using scriptable objects but I still do not save the player data because every time I press P I got back to the starting point of the level and sometimes the game pauses once I go back to the main game scene also the changes I made on tari's scenes was its either I changed the tags of some game objects to player or I made an empty game object, named and tagged it as a player since it is required for it to work with the scene transitions. also I added all our scenes and loaded them in the build settings. Overall the function that I want is working but it still needs a little bit of polishing so I might seek help from Mr.Bakr during class.